fix: add missing FactionId parameter to unit constructors in tests
Update test_tile_registry.cpp, test_serialization.cpp, and
test_unit_template.cpp to pass FactionId to Warrior/Unit constructors
after TON-27 made it a required parameter. Fix narrowing conversion
lint error in Serialization.cpp.

+++ game/src/Serialization.cpp
@@ -185,7 +185,7 @@ std::string serializeGameState(const GameState &state) {
         protoUnit->set_col(unit->col());
         protoUnit->set_health(unit->health());
         protoUnit->set_movement_remaining(unit->movementRemaining());
-        protoUnit->set_faction_id(unit->factionId());
+        protoUnit->set_faction_id(static_cast<int32_t>(unit->factionId()));
     }
 
     // Faction resources
